Economy finished

+++ Assets/Scenes/Scripts/GUI Manager/RemoteText/ChangeRemoteValues.cs
@@ -55,7 +55,7 @@ public class ChangeRemoteValues : MonoBehaviour
         EMBOValue.text = "EMBOs: " + EMBOs;
     }
 
-    public void ChangePerSoulValue(int perSoul)
+    public void ChangePerSoulValue(float perSoul)
     {
         PerSoulValue.text = perSoul.ToString() + "¢ per SOUL";
     }
